backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.4
Flask-CORS==4.0.0
python-socketio==5.9.0
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0